refactor: 주입 정보를 관리하는 객체와 주입해주는 객체 분리
