+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -5,14 +5,24 @@ import woowacourse.shopping.data.CartProductDao
 import woowacourse.shopping.data.ShoppingDatabase
 
 class ShoppingApplication : Application() {
-    lateinit var appContainer: AppContainer
+    private lateinit var appContainer: AppContainer
+    lateinit var injector: Injector
         private set
 
     override fun onCreate() {
         super.onCreate()
+        setupContainer()
+        setupInjector()
+    }
+
+    private fun setupContainer() {
         appContainer = AppContainer()
         appContainer.registerProviders {
             provider(CartProductDao::class to ShoppingDatabase.getInstance(applicationContext)::cartProductDao)
         }
+    }
+
+    private fun setupInjector() {
+        injector = Injector(appContainer)
     }
 }
