+++ app/src/main/java/woowacourse/shopping/ui/ViewModelFactory.kt
@@ -9,7 +9,7 @@ import woowacourse.shopping.ShoppingApplication
 val viewModelFactory = object : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>, extras: CreationExtras): T {
         val application = checkNotNull(extras[APPLICATION_KEY]) as ShoppingApplication
-        val appContainer = application.appContainer
-        return appContainer.inject(modelClass.kotlin)
+        val injector = application.injector
+        return injector.inject(modelClass.kotlin)
     }
 }
